Make the "Go Back" button on upload warning/error work even for no-JS users
The uploadError() method is called at least when trying to upload a file of an unsupported type, e.g. .svg or .tiff.
Now the "Go Back" button correctly takes the user back to the main upload form even without JS being enabled.

Note/TODO: the "action=submit" part in the URL is responsible for showing what looks to be the "sessionfailure" i18n msg.
I'm not sure why it's doing that but I decided not to investigate that since it was already there.

Bug: T248390
Change-Id: Iefd0dc2b62705b18316f0cd703422e6167b0fc07