Fix the [agree] links on network pages (Special:FanHome w/ appropriate sport_id and team_id URL params set)
Previously the link would get followed, even though that's primarily meant for no-JS users (and even *that* isn't working correctly currently, the URLs are lacking "vote=1" bit which is set for the [agree] links on user profile pages but not here).
Now users w/ JS enabled don't follow the link and the AJAX request fires successfully, as intended.

Change-Id: Iafe0c8cb2783b36ac0111d0e134fdf2490a2d294